+++ tests/units/classes/mock/generator.php
@@ -1886,6 +1886,11 @@ class generator extends atoum\test
 	{
 		$this
 			->if($generator = new testedClass())
+			->and($reflectionTypeController = new mock\controller())
+			->and($reflectionTypeController->__construct = function() {})
+			->and($reflectionTypeController->isBuiltin = true)
+			->and($reflectionTypeController->__toString = $returnType = 'string')
+			->and($reflectionType = new \mock\reflectionType())
 			->and($reflectionMethodController = new mock\controller())
 			->and($reflectionMethodController->__construct = function() {})
 			->and($reflectionMethodController->getName = $methodName = uniqid())
@@ -1899,7 +1904,7 @@ class generator extends atoum\test
 			->and($reflectionMethodController->isAbstract = false)
 			->and($reflectionMethodController->returnsReference = false)
 			->and($reflectionMethodController->hasReturnType = true)
-			->and($reflectionMethodController->getReturnType = 'string')
+			->and($reflectionMethodController->getReturnType = $reflectionType)
 			->and($reflectionMethod = new \mock\reflectionMethod(null, null))
 			->and($reflectionClassController = new mock\controller())
 			->and($reflectionClassController->__construct = function() {})
@@ -1935,7 +1940,7 @@ class generator extends atoum\test
 					"\t\t\t" . '$this->getMockController()->invoke(\'__construct\', func_get_args());' . PHP_EOL .
 					"\t\t" . '}' . PHP_EOL .
 					"\t" . '}' . PHP_EOL .
-					"\t" . 'public function ' . $methodName . '(): string' . PHP_EOL .
+					"\t" . 'public function ' . $methodName . '(): ' . $returnType . PHP_EOL .
 					"\t" . '{' . PHP_EOL .
 					"\t\t" . '$arguments = array_merge(array(), array_slice(func_get_args(), 0));' . PHP_EOL .
 					"\t\t" . 'if (isset($this->getMockController()->' . $methodName . ') === true)' . PHP_EOL .
@@ -1957,6 +1962,91 @@ class generator extends atoum\test
 					'}' . PHP_EOL .
 					'}'
 				)
+		;
+	}
+
+
+	/** @php >= 7.0 */
+	public function testGetMockedClassCodeForInterfaceWithReturnTypeNotBuiltIn()
+	{
+		$this
+			->if($generator = new testedClass())
+			->and($reflectionTypeController = new mock\controller())
+			->and($reflectionTypeController->__construct = function() {})
+			->and($reflectionTypeController->isBuiltin = false)
+			->and($reflectionTypeController->__toString = $returnType = 'Mock\Foo')
+			->and($reflectionType = new \mock\reflectionType())
+			->and($reflectionMethodController = new mock\controller())
+			->and($reflectionMethodController->__construct = function() {})
+			->and($reflectionMethodController->getName = $methodName = uniqid())
+			->and($reflectionMethodController->isConstructor = false)
+			->and($reflectionMethodController->getParameters = array())
+			->and($reflectionMethodController->isPublic = true)
+			->and($reflectionMethodController->isProtected = false)
+			->and($reflectionMethodController->isPrivate = false)
+			->and($reflectionMethodController->isFinal = false)
+			->and($reflectionMethodController->isStatic = false)
+			->and($reflectionMethodController->isAbstract = false)
+			->and($reflectionMethodController->returnsReference = false)
+			->and($reflectionMethodController->hasReturnType = true)
+			->and($reflectionMethodController->getReturnType = $reflectionType)
+			->and($reflectionMethod = new \mock\reflectionMethod(null, null))
+			->and($reflectionClassController = new mock\controller())
+			->and($reflectionClassController->__construct = function() {})
+			->and($reflectionClassController->getName = function() use (& $realClass) { return $realClass; })
+			->and($reflectionClassController->isFinal = false)
+			->and($reflectionClassController->isInterface = false)
+			->and($reflectionClassController->getMethods = array($reflectionMethod))
+			->and($reflectionClassController->getConstructor = null)
+			->and($reflectionClassController->isAbstract = false)
+			->and($reflectionClass = new \mock\reflectionClass(null))
+			->and($generator->setReflectionClassFactory(function() use ($reflectionClass) { return $reflectionClass; }))
+			->and($adapter = new atoum\test\adapter())
+			->and($adapter->class_exists = function($class) use (& $realClass) { return ($class == '\\' . $realClass); })
+			->and($generator->setAdapter($adapter))
+			->then
+			->string($generator->getMockedClassCode($realClass = uniqid()))->isEqualTo(
+				'namespace mock {' . PHP_EOL .
+				'final class ' . $realClass . ' extends \\' . $realClass . ' implements \mageekguy\atoum\mock\aggregator' . PHP_EOL .
+				'{' . PHP_EOL .
+				$this->getMockControllerMethods() .
+				"\t" . 'public function __construct(\mageekguy\atoum\mock\controller $mockController = null)' . PHP_EOL .
+				"\t" . '{' . PHP_EOL .
+				"\t\t" . 'if ($mockController === null)' . PHP_EOL .
+				"\t\t" . '{' . PHP_EOL .
+				"\t\t\t" . '$mockController = \mageekguy\atoum\mock\controller::get();' . PHP_EOL .
+				"\t\t" . '}' . PHP_EOL .
+				"\t\t" . 'if ($mockController !== null)' . PHP_EOL .
+				"\t\t" . '{' . PHP_EOL .
+				"\t\t\t" . '$this->setMockController($mockController);' . PHP_EOL .
+				"\t\t" . '}' . PHP_EOL .
+				"\t\t" . 'if (isset($this->getMockController()->__construct) === true)' . PHP_EOL .
+				"\t\t" . '{' . PHP_EOL .
+				"\t\t\t" . '$this->getMockController()->invoke(\'__construct\', func_get_args());' . PHP_EOL .
+				"\t\t" . '}' . PHP_EOL .
+				"\t" . '}' . PHP_EOL .
+				"\t" . 'public function ' . $methodName . '(): \\' . $returnType . PHP_EOL .
+				"\t" . '{' . PHP_EOL .
+				"\t\t" . '$arguments = array_merge(array(), array_slice(func_get_args(), 0));' . PHP_EOL .
+				"\t\t" . 'if (isset($this->getMockController()->' . $methodName . ') === true)' . PHP_EOL .
+				"\t\t" . '{' . PHP_EOL .
+				"\t\t\t" . '$return = $this->getMockController()->invoke(\'' . $methodName . '\', $arguments);' . PHP_EOL .
+				"\t\t\t" . 'return $return;' . PHP_EOL .
+				"\t\t" . '}' . PHP_EOL .
+				"\t\t" . 'else' . PHP_EOL .
+				"\t\t" . '{' . PHP_EOL .
+				"\t\t\t" . '$this->getMockController()->addCall(\'' . $methodName . '\', $arguments);' . PHP_EOL .
+				"\t\t\t" . '$return = call_user_func_array(\'parent::' . $methodName . '\', $arguments);' . PHP_EOL .
+				"\t\t\t" . 'return $return;' . PHP_EOL .
+				"\t\t" . '}' . PHP_EOL .
+				"\t" . '}' . PHP_EOL .
+				"\t" . 'public static function getMockedMethods()' . PHP_EOL .
+				"\t" . '{' . PHP_EOL .
+				"\t\t" . 'return ' . var_export(array('__construct', $methodName), true) . ';' . PHP_EOL .
+				"\t" . '}' . PHP_EOL .
+				'}' . PHP_EOL .
+				'}'
+			)
 		;
 	}
 
@@ -3200,6 +3290,11 @@ class generator extends atoum\test
 	{
 		$this
 			->if($generator = new testedClass())
+			->and($reflectionTypeController = new mock\controller())
+			->and($reflectionTypeController->__construct = function() {})
+			->and($reflectionTypeController->isBuiltin = true)
+			->and($reflectionTypeController->__toString = $returnType = 'string')
+			->and($reflectionType = new \mock\reflectionType())
 			->and($reflectionMethodController = new mock\controller())
 			->and($reflectionMethodController->__construct = function() {})
 			->and($reflectionMethodController->getName = $methodName = uniqid())
@@ -3213,7 +3308,7 @@ class generator extends atoum\test
 			->and($reflectionMethodController->isAbstract = false)
 			->and($reflectionMethodController->returnsReference = false)
 			->and($reflectionMethodController->hasReturnType = true)
-			->and($reflectionMethodController->getReturnType = 'string')
+			->and($reflectionMethodController->getReturnType = $reflectionType)
 			->and($reflectionMethod = new \mock\reflectionMethod(null, null))
 			->and($reflectionClassController = new mock\controller())
 			->and($reflectionClassController->__construct = function() {})
@@ -3249,7 +3344,7 @@ class generator extends atoum\test
 					"\t\t\t" . '$this->getMockController()->invoke(\'__construct\', func_get_args());' . PHP_EOL .
 					"\t\t" . '}' . PHP_EOL .
 					"\t" . '}' . PHP_EOL .
-					"\t" . 'public function ' . $methodName . '(): string' . PHP_EOL .
+					"\t" . 'public function ' . $methodName . '(): ' . $returnType . PHP_EOL .
 					"\t" . '{' . PHP_EOL .
 					"\t\t" . '$arguments = array_merge(array(), array_slice(func_get_args(), 0));' . PHP_EOL .
 					"\t\t" . 'if (isset($this->getMockController()->' . $methodName . ') === true)' . PHP_EOL .